Call connectionIssue if unable to connect
Previously, connectionIssue was only called if a response was
received from the memcached service, and the response was a
SERVER_ERROR.  However, if the node client could not connect to
a memcached server at all, this case was not handled. Now, on the
error event on the stream, call connectionIssue.

Additionally, before trying to connect to memcached, first check
whether any healthy servers exist.