Ensure that Client instantiated with 'new'
Add check for when 'this' is undefined which (in strict mode)
signals that the user called the function directly instead of
using the 'new' operator. In this case, return a newly
instantiated client using the same parameters.

This prevents indescriptive errors when applying the global
config to the newly instantiated client when the user forgets
to instantiate the client with 'new' operator.